[105] Refactored the importing of govuk frontend
