+++ app/components/base.scss
@@ -1,0 +1,2 @@
+@import "../webpacker/styles/govuk_style_setup";
+@import "~govuk-frontend/govuk/base";
