+++ app/webpacker/packs/application.js
@@ -1,6 +1,6 @@
-import './govuk_assets_import'
+import '../scripts/govuk_assets_import'
 import '../styles/application.scss'
-import './components'
+import '../scripts/components'
 import { initAll } from 'govuk-frontend'
 
 initAll()
